refactor: reuse DA_ADMIN and HLX_ADMIN constants from nx/utils/utils.js
