+++ nx/blocks/media-library/core/paths.js
@@ -1,5 +1,6 @@
 import { daFetch } from '../../../utils/daFetch.js';
-import { Paths, Domains, DA_LIVE_EDIT_BASE, DA_ADMIN } from './constants.js';
+import { DA_ADMIN } from '../../../utils/utils.js';
+import { Paths, Domains, DA_LIVE_EDIT_BASE } from './constants.js';
 import { ErrorCodes, logMediaLibraryError } from './errors.js';
 import { t } from './messages.js';
 import {
